# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Change tracker api docs",
    "description": "",
    "license": {
      "name": "MIT",
      "url": ""
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/api": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Send default response to api endpoint",
        "description": "",
        "operationId": "api",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
    "/api/configs": {
      "get": {
        "tags": [
          "configs"
        ],
        "summary": "Get configs available",
        "description": "Get available configs or config of a given url. e.g. https://wildbit.com/jobs",
        "operationId": "configs",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "URL that need its data.",
          "required": false,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/AllConfigs"
            }
          }
        }
      }
    },
    "/api/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get config rules available",
        "description": "Get available config rules or single rule of a url. e.g https://wildbit.com/jobs",
        "operationId": "rules",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "URL that need its data.",
          "required": false,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Config"
            }
          }
        }
      }
    },
    "/api/status": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Get changesets of website",
        "description": "Get changesets of a given url e.g. https://wildbit.com/jobs",
        "operationId": "status",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [{
          "name": "url",
          "in": "query",
          "description": "URL that need its data.",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Config"
            }
          },
          "400": {
            "description": "URL is invalid or missing",
            "schema": {
              "$ref":"#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get changesets of all website",
        "description": "Get changesets of all websuites stored",
        "operationId": "jobs",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/AllConfigs"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "DefaultResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "Frame": {
      "type": "object",
      "properties": {
        "index": {
          "type": "number"
        },
        "excludes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "includes": {
          "type": "array",
          "items": {
            "type": "object",
            "proerties": {
              "expr": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "Selections": {
      "type": "object",
      "properties": {
        "frames": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Frame"
          }
        },
        "dynamic": {
          "type": "boolean"
        },
        "delay": {
          "type": "number"
        }
      }
    },
    "Config": {
      "type": "object",
      "properties": {
        "includeStyle": {
          "type": "boolean"
        },
        "selections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Selections"
          }
        },
        "ignoreEmptyText": {
          "type": "boolean"
        },
        "dataAttr": {
          "type": "string"
        }
      }
    },
    "WebInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "uri": {
          "type": "string"
        },
        "config": {
          "type": "object",
          "$ref": "#/definitions/Config"
        },
        "tags": {
          "type": "string"
        },
        "content_type": {
          "type": "number"
        },
        "state": {
          "type": "number"
        },
        "schedule": {
          "type": "object"
        },
        "ts": {
          "type": "string"
        }
      }
    },
    "AllConfigs": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/WebInfo"
      }
    }
  }
}